% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeesApprox.r
\name{calcprobR}
\alias{calcprobR}
\title{Calculate Probability for Length Classes}
\usage{
calcprobR(tempLens, tempNs, xout, LenBins)
}
\arguments{
\item{tempLens}{Vector of Lengths}

\item{tempNs}{Vector of numbers for each length in tempLens}

\item{xout}{Points to interpolate at}

\item{LenBins}{Length bins}
}
\value{
Probability for each length bin
}
\description{
Calculate Probability for Length Classes
}
