% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeesApprox.r
\name{dnormal}
\alias{dnormal}
\title{Double-normal selectivity function}
\usage{
dnormal(lens, lfs, sl, sr)
}
\arguments{
\item{lens}{Vector of lengths}

\item{lfs}{Length at full selectivity}

\item{sl}{Slope of left curve}

\item{sr}{Slope of right curve}
}
\value{
Selectivity vector for \code{lens}
}
\description{
Double-normal selectivity function
}
