% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeesApprox.r
\name{LeesApproxR}
\alias{LeesApproxR}
\title{R version of approximation method for Lee's Phenomenon}
\usage{
LeesApproxR(
  FVec,
  ngtg = 1001,
  maxsd,
  binwidth,
  M,
  Linf,
  K,
  t0,
  LFS,
  L5,
  Vmaxlen,
  LinfCV,
  maxage
)
}
\arguments{
\item{FVec}{Annual Fs}

\item{ngtg}{Number of growth-type-groups}

\item{maxsd}{Maximum number of standard deviations for length-at-age dist}

\item{binwidth}{Width of length bins}

\item{M}{Natural mortality}

\item{Linf}{Asymptotic length}

\item{K}{von B K parameter}

\item{t0}{von B t0}

\item{LFS}{Length at full selection}

\item{L5}{First length at 5\% selection}

\item{Vmaxlen}{Vulnerability at maximum length}

\item{LinfCV}{CV of length-at-age dist}

\item{maxage}{maximum age}
}
\value{
A list
}
\description{
R version of approximation method for Lee's Phenomenon
}
